Move new PathAttribute, part deux: The unsaved vim buffers
