+++ src/bgp/message/update.rs
@@ -1,5 +1,4 @@
 use crate::bgp::message::Header;
-use bytes::BytesMut;
 use octseq::{Octets, Parser};
 use log::{debug, error, warn};
 
@@ -14,8 +13,7 @@ use crate::bgp::message::nlri::{
     RouteTargetNlri
 };
 
-use core::iter::Peekable;
-use std::net::{IpAddr, Ipv4Addr, Ipv6Addr};
+use std::net::Ipv4Addr;
 use std::fmt;
 use crate::util::parser::{parse_ipv4addr, ParseError};
 
@@ -2039,870 +2037,6 @@ impl LargeCommunity {
 }
 
 
-//------------ UpdateBuilder -------------------------------------------------
-
-use octseq::{FreezeBuilder, OctetsBuilder, ShortBuf};
-use crate::bgp::message::attr_change_set::AttrChangeSet;
-use crate::bgp::message::MsgType;
-use crate::bgp::message::update_builder::{
-    MpReachNlriBuilder,
-    MpUnreachNlriBuilder,
-    StandardCommunitiesBuilder
-};
-
-//use rotonda_fsm::bgp::session::AgreedConfig;
-
-use crate::bgp::path_attributes as new_pas;
-#[derive(Debug)]
-pub struct UpdateBuilder<Target> {
-    target: Target,
-    //config: AgreedConfig, //FIXME this lives in rotonda_fsm, but that
-    //depends on routecore. Sounds like a depedency hell.
-    announcements: Vec<Nlri<Vec<u8>>>,
-    announcements_len: usize,
-    withdrawals: Vec<Nlri<Vec<u8>>>,
-    withdrawals_len: usize,
-    addpath_enabled: Option<bool>, // for conventional nlri (unicast v4)
-    //attributes: Vec<PathAttribute<'a, Vec<u8>>>, // XXX this lifetime is..
-                                                    //not nice
-    // attributes:
-    origin: Option<new_pas::Origin>,
-    aspath: Option<AsPath<Vec<u8>>>,
-    nexthop: Option<new_pas::NextHop>,
-    multi_exit_disc: Option<new_pas::MultiExitDisc>,
-    local_pref: Option<new_pas::LocalPref>,
-
-    standard_communities_builder: Option<StandardCommunitiesBuilder>,
-
-    attributes_len: usize,
-    total_pdu_len: usize,
-
-    // MP_REACH_NLRI and MP_UNREACH_NLRI can only occur once (like any path
-    // attribute), and can carry only a single tuple of (AFI, SAFI).
-    // My interpretation of RFC4760 means one can mix conventional
-    // NLRI/withdrawals (so, v4 unicast) with one other (AFI, SAFI) in an
-    // MP_(UN)REACH_NLRI path attribute.
-    // Question is: can one also put (v4, unicast) in an MP_* attribute, and,
-    // then also in the conventional part (at the end of the PDU)? 
-    mp_reach_nlri_builder: Option<MpReachNlriBuilder>,
-    mp_unreach_nlri_builder: Option<MpUnreachNlriBuilder>,
-}
-
-impl<Target: OctetsBuilder> UpdateBuilder<Target> {
-
-    const MAX_PDU: usize = 4096; // XXX should come from NegotiatedConfig
-
-    pub fn from_target(mut target: Target) -> Result<Self, ShortBuf> {
-        //target.truncate(0);
-        let mut h = Header::<&[u8]>::new();
-        h.set_length(19 + 2 + 2);
-        h.set_type(MsgType::Update);
-        let _ =target.append_slice(h.as_ref());
-
-        Ok(UpdateBuilder {
-            target,
-            announcements: Vec::new(),
-            announcements_len: 0,
-            withdrawals: Vec::new(),
-            withdrawals_len: 0,
-            addpath_enabled: None,
-
-            //attributes:
-            origin: None,
-            aspath: None,
-            nexthop: None,
-            multi_exit_disc: None,
-            local_pref: None,
-
-            standard_communities_builder: None,
-
-            attributes_len: 0,
-            total_pdu_len: 19 + 2 + 2,
-            mp_reach_nlri_builder: None,
-            mp_unreach_nlri_builder: None,
-        })
-    }
-
-    //--- Withdrawals
-
-    pub fn add_withdrawal(&mut self, withdrawal: &Nlri<Vec<u8>>)
-        -> Result<(), ComposeError>
-    {
-        //println!("add_withdrawal for {}", withdrawal.prefix().unwrap());
-        match *withdrawal {
-            Nlri::Unicast(b) => {
-                if b.is_v4() {
-                    if let Some(addpath_enabled) = self.addpath_enabled {
-                        if addpath_enabled != b.is_addpath() {
-                            return Err(ComposeError::IllegalCombination)
-                        }
-                    } else {
-                        self.addpath_enabled = Some(b.is_addpath());
-                    }
-
-                    let new_bytes_num = withdrawal.compose_len();
-                    let new_total = self.total_pdu_len + new_bytes_num;
-                    if new_total > Self::MAX_PDU {
-                        return Err(ComposeError::PduTooLarge(new_total));
-                    }
-                    self.withdrawals.push(withdrawal.clone());
-                    self.withdrawals_len += new_bytes_num;
-                    self.total_pdu_len = new_total;
-                } else {
-                    // Nlri::Unicast only holds IPv4 and IPv6, so this must be
-                    // IPv6.
-                    let new_bytes_num = match self.mp_unreach_nlri_builder {
-                        None => {
-                            let builder = MpUnreachNlriBuilder::new(
-                                AFI::Ipv6, SAFI::Unicast,
-                                b.is_addpath()
-                            );
-                            let res = builder.compose_len(withdrawal);
-                            self.mp_unreach_nlri_builder = Some(builder);
-                            res
-                            
-                        }
-                        Some(ref builder) => {
-                            if !builder.valid_combination(
-                                AFI::Ipv6, SAFI::Unicast, b.is_addpath()
-                            ) {
-                                // We are already constructing a
-                                // MP_UNREACH_NLRI but for a different
-                                // AFI,SAFI than the prefix in `withdrawal`,
-                                // or we are mixing addpath with non-addpath.
-                                return Err(ComposeError::IllegalCombination);
-                            }
-                            builder.compose_len(withdrawal)
-                        }
-                    };
-
-                    let new_total = self.total_pdu_len + new_bytes_num;
-
-                    if new_total > Self::MAX_PDU {
-                        return Err(ComposeError::PduTooLarge(new_total));
-                    }
-
-                    if let Some(ref mut builder) = self.mp_unreach_nlri_builder {
-                        builder.add_withdrawal(withdrawal);
-                        self.attributes_len += new_bytes_num;
-                        self.total_pdu_len = new_total;
-                    } else {
-                        // We always have Some builder at this point.
-                        unreachable!()
-                    }
-                }
-
-            }
-            _ => todo!() // TODO use the MpUnreachNlriBuilder for these
-        };
-        Ok(())
-    }
-
-    pub fn withdrawals_from_iter<I>(&mut self, withdrawals: &mut Peekable<I>)
-        -> Result<(), ComposeError>
-    where I: Iterator<Item = Nlri<Vec<u8>>>
-    {
-        while let Some(w) = withdrawals.peek() {
-            match self.add_withdrawal(w) {
-                Ok(_) => { withdrawals.next(); }
-                Err(e) => return Err(e)
-            }
-        }
-        Ok(())
-    }
-
-    pub fn append_withdrawals(&mut self, withdrawals: &mut Vec<Nlri<Vec<u8>>>)
-        -> Result<(), ComposeError>
-    {
-        for (idx, w) in withdrawals.iter().enumerate() {
-            if let Err(e) = self.add_withdrawal(w) {
-
-                *withdrawals = withdrawals[idx..].to_vec();
-                return Err(e);
-            }
-        }
-        Ok(())
-    }
-
-    //--- Path Attributes
-
-    pub fn set_origin(&mut self, origin: OriginType)
-        -> Result<(), ComposeError>
-    {
-        if self.origin.is_none() {
-            // Check if this goes over the total PDU len. That will happen
-            // before we will go over the u16 for Total Path Attributes Length
-            // so we don't have to check for that.
-            let new_total = self.total_pdu_len + 4; // XXX should this 4 come
-                                                    // from a
-                                                    // `fn compose_len()` on
-                                                    // the attribute itself
-                                                    // perhaps?
-            if new_total > Self::MAX_PDU {
-                return Err(ComposeError::PduTooLarge(new_total));
-            } else {
-                self.total_pdu_len = new_total;
-                self.attributes_len += 4
-            }
-        }
-        self.origin = Some(new_pas::Origin::new(origin));
-        Ok(())
-    }
-
-    pub fn set_aspath<Octs>(&mut self , aspath: AsPath<Vec<u8>>)
-        -> Result<(), ComposeError>
-    {
-        if aspath.compose_len() > u16::MAX.into()  {
-            return Err(ComposeError::AttributeTooLarge(
-                 PathAttributeType::AsPath, aspath.compose_len()
-            ));
-        }
-        if let Some(old_aspath) = &self.aspath {
-            let new_total = self.total_pdu_len
-                - old_aspath.compose_len()
-                + aspath.compose_len();
-            if new_total > Self::MAX_PDU {
-                return Err(ComposeError::PduTooLarge(new_total));
-            } else {
-                self.total_pdu_len -= old_aspath.compose_len();
-                self.attributes_len -= old_aspath.compose_len();
-            }
-        }
-        self.attributes_len += aspath.compose_len();
-        self.total_pdu_len += aspath.compose_len();
-        self.aspath = Some(aspath);
-        Ok(())
-    }
-
-    pub fn set_nexthop(&mut self, addr: IpAddr) -> Result<(), ComposeError> {
-        // Depending on the variant of `addr` we add/update either:
-        // - the conventional NEXT_HOP path attribute (IPv4); or
-        // - we update/create a MpReachNlriBuilder (IPv6)
-        match addr {
-            IpAddr::V4(a) => {
-                if self.nexthop.is_none() {
-                    // NEXT_HOP path attribute is 7 bytes long.
-                    let new_total = self.total_pdu_len + 7;
-                    if new_total > Self::MAX_PDU {
-                        return Err(ComposeError::PduTooLarge(new_total));
-                    } else {
-                        self.total_pdu_len = new_total;
-                        self.attributes_len += 7
-                    }
-                    self.nexthop = Some(new_pas::NextHop::new(a));
-                }
-            }
-            IpAddr::V6(a) => {
-                if let Some(ref mut builder) = self.mp_reach_nlri_builder {
-                    // Given that we have a builder, there is already either a
-                    // NextHop::Ipv6 or a NextHop::Ipv6LL. Setting the
-                    // non-link-local address does not change the length of
-                    // the next hop part, so there is no need to update the
-                    // total_pdu_len and the attributes_len.
-                    builder.set_nexthop(a);
-                } else {
-                    let builder = MpReachNlriBuilder::new(
-                        AFI::Ipv6, SAFI::Unicast, NextHop::Ipv6(a), false
-                        );
-                    let new_bytes = builder.compose_len_empty();
-                    let new_total = self.total_pdu_len + new_bytes;
-                    if new_total > Self::MAX_PDU {
-                        return Err(ComposeError::PduTooLarge(new_total));
-                    }
-                    self.attributes_len += new_bytes;
-                    self.total_pdu_len = new_total;
-                    self.mp_reach_nlri_builder = Some(builder);
-                }
-            }
-        }
-        Ok(())
-    }
-
-    pub fn set_nexthop_ll(&mut self, addr: Ipv6Addr)
-        -> Result<(), ComposeError>
-    {
-        // XXX We could/should check for addr.is_unicast_link_local() once
-        // that lands in stable.
-        
-        if let Some(ref mut builder) = self.mp_reach_nlri_builder {
-            let new_bytes = match builder.get_nexthop() {
-                NextHop::Ipv6(_) => 16,
-                NextHop::Ipv6LL(_,_) => 0,
-                _ => unreachable!()
-            };
-
-            let new_total = self.total_pdu_len + new_bytes;
-            if new_total > Self::MAX_PDU {
-                return Err(ComposeError::PduTooLarge(new_total));
-            }
-            builder.set_nexthop_ll(addr);
-            self.attributes_len += new_bytes;
-            self.total_pdu_len = new_total;
-        } else {
-            let builder = MpReachNlriBuilder::new(
-                AFI::Ipv6, SAFI::Unicast, NextHop::Ipv6LL(0.into(), addr),
-                false
-            );
-            let new_bytes = builder.compose_len_empty();
-            let new_total = self.total_pdu_len + new_bytes;
-            if new_total > Self::MAX_PDU {
-                return Err(ComposeError::PduTooLarge(new_total));
-            }
-            self.attributes_len += new_bytes;
-            self.total_pdu_len = new_total;
-            self.mp_reach_nlri_builder = Some(builder);
-        }
-
-        Ok(())
-    }
-
-    pub fn set_multi_exit_disc(&mut self, med: new_pas::MultiExitDisc)
-    -> Result<(), ComposeError>
-    {
-        let new_bytes = med.compose_len();
-        let new_total = self.total_pdu_len + new_bytes;
-        if new_total > Self::MAX_PDU {
-            return Err(ComposeError::PduTooLarge(new_total));
-        }
-        self.multi_exit_disc = Some(med);
-        self.total_pdu_len = new_total;
-        self.attributes_len += new_bytes;
-        Ok(())
-    }
-
-    pub fn set_local_pref(&mut self, local_pref: new_pas::LocalPref)
-    -> Result<(), ComposeError>
-    {
-        let new_bytes = local_pref.compose_len();
-        let new_total = self.total_pdu_len + new_bytes;
-        if new_total > Self::MAX_PDU {
-            return Err(ComposeError::PduTooLarge(new_total));
-        }
-        self.local_pref = Some(local_pref);
-        self.total_pdu_len = new_total;
-        self.attributes_len += new_bytes;
-        Ok(())
-    }
-
-
-    //--- Announcements
-
-    pub fn add_announcement(&mut self, announcement: &Nlri<Vec<u8>>)
-        -> Result<(), ComposeError>
-    {
-        match *announcement {
-            Nlri::Unicast(b) => {
-                if b.is_v4() {
-                    if let Some(addpath_enabled) = self.addpath_enabled {
-                        if addpath_enabled != b.is_addpath() {
-                            return Err(ComposeError::IllegalCombination)
-                        }
-                    } else {
-                        self.addpath_enabled = Some(b.is_addpath());
-                    }
-
-                    let new_bytes_num = announcement.compose_len();
-                    let new_total = self.total_pdu_len + new_bytes_num;
-                    if new_total > Self::MAX_PDU {
-                        return Err(ComposeError::PduTooLarge(new_total));
-                    }
-                    self.announcements.push(announcement.clone());
-                    self.announcements_len += new_bytes_num;
-                    self.total_pdu_len = new_total;
-                    
-                } else {
-                    // Nlri::Unicast only holds IPv4 and IPv6, so this must be
-                    // IPv6 unicast.
-
-                    let new_bytes_num = match self.mp_reach_nlri_builder {
-                        None => {
-                            let nexthop = NextHop::Ipv6(0.into());
-                            let builder = MpReachNlriBuilder::new(
-                                AFI::Ipv6, SAFI::Unicast,
-                                nexthop,
-                                b.is_addpath()
-                            );
-                            let res = builder.compose_len_empty() + 
-                                builder.compose_len(announcement);
-                            self.mp_reach_nlri_builder = Some(builder);
-                            res
-                            
-                        }
-                        Some(ref builder) => {
-                            //TODO we should allow multicast here, but then we
-                            //should also check whether a prefix is_multicast.
-                            //Similarly, should we check for is_unicast?
-                            //Or should we handle anything other than unicast
-                            //in the outer match anyway, as that is never a
-                            //conventional announcement anyway?
-                            
-                            if !builder.valid_combination(
-                                AFI::Ipv6, SAFI::Unicast, b.is_addpath()
-                            ) {
-                                return Err(ComposeError::IllegalCombination);
-                            }
-
-                            builder.compose_len(announcement)
-                        }
-                    };
-
-                    let new_total = self.total_pdu_len + new_bytes_num;
-
-                    if new_total > Self::MAX_PDU {
-                        return Err(ComposeError::PduTooLarge(new_total));
-                    }
-
-                    if let Some(ref mut builder) = self.mp_reach_nlri_builder {
-                        builder.add_announcement(announcement);
-                        self.attributes_len += new_bytes_num;
-                        self.total_pdu_len = new_total;
-                    } else {
-                        // We always have Some builder at this point.
-                        unreachable!()
-                    }
-                }
-            }
-            _ => todo!() // TODO use MpReachNlriBuilder once we have that.
-        }
-
-        Ok(())
-    }
-
-    pub fn announcements_from_iter<I>(
-        &mut self, announcements: &mut Peekable<I>
-    ) -> Result<(), ComposeError>
-        where I: Iterator<Item = Nlri<Vec<u8>>>
-    {
-        while let Some(a) = announcements.peek() {
-            match self.add_announcement(a) {
-                Ok(_) => { announcements.next(); }
-                Err(e) => return Err(e)
-            }
-        }
-        Ok(())
-    }
-
-    //--- Standard communities
-    
-    pub fn add_community(&mut self, community: StandardCommunity)
-        -> Result<(), ComposeError>
-    {
-
-        if let Some(ref mut builder) = self.standard_communities_builder {
-            let new_bytes = builder.compose_len(community);
-            let new_total = self.total_pdu_len + new_bytes;
-            if new_total > Self::MAX_PDU {
-                return Err(ComposeError::PduTooLarge(new_total));
-            }
-            builder.add_community(community);
-            self.attributes_len += new_bytes;
-            self.total_pdu_len = new_total;
-        } else {
-            let mut builder = StandardCommunitiesBuilder::new();
-            let new_bytes = builder.compose_len_empty()
-                + builder.compose_len(community); 
-            let new_total = self.total_pdu_len + new_bytes;
-            if new_total > Self::MAX_PDU {
-                return Err(ComposeError::PduTooLarge(new_total));
-            }
-            builder.add_community(community);
-            self.standard_communities_builder = Some(builder);
-            self.attributes_len += new_bytes;
-            self.total_pdu_len = new_total;
-        }
-
-        Ok(())
-    }
-
-}
-
-#[derive(Debug)]
-pub enum ComposeError{
-    /// Exceeded maximum PDU size, data field carries the violating length.
-    PduTooLarge(usize),
-
-    // TODO proper docstrings, first see how/if we actually use these.
-    AttributeTooLarge(PathAttributeType, usize),
-    AttributesTooLarge(usize),
-    IllegalCombination,
-    EmptyMpReachNlri,
-    EmptyMpUnreachNlri,
-    WrongAddressType,
-
-    InvalidAttribute,
-
-    /// Variant for `octseq::builder::ShortBuf`
-    ShortBuf,
-    /// Wrapper for util::parser::ParseError, used in `fn into_message`
-    ParseError(ParseError)
-}
-
-impl From<ShortBuf> for ComposeError {
-    fn from(_: ShortBuf) -> ComposeError {
-        ComposeError::ShortBuf
-    }
-}
-impl From<ParseError> for ComposeError {
-    fn from(pe: ParseError) -> ComposeError {
-        ComposeError::ParseError(pe)
-    }
-}
-
-
-impl std::error::Error for ComposeError { }
-impl fmt::Display for ComposeError {
-    fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
-        match self {
-            ComposeError::PduTooLarge(n) => {
-                write!(f, "oversized PDU: {n} bytes")
-            }
-            ComposeError::AttributeTooLarge(attr, n) => {
-                write!(f, "oversized attribute {attr}: {n} bytes")
-            }
-            ComposeError::AttributesTooLarge(n) => {
-                write!(f, "total path attributes too large: {n} bytes")
-            }
-            ComposeError::IllegalCombination => {
-                write!(f, "illegal combination of prefixes/attributes")
-            }
-            ComposeError::EmptyMpReachNlri => {
-                write!(f, "missing NLRI in MP_REACH_NLRI")
-            }
-            ComposeError::EmptyMpUnreachNlri => {
-                write!(f, "missing NLRI in MP_UNREACH_NLRI")
-            }
-            ComposeError::WrongAddressType => {
-                write!(f, "wrong address type")
-            }
-            ComposeError::InvalidAttribute => {
-                write!(f, "invalid attribute")
-            }
-            ComposeError::ShortBuf => {
-                ShortBuf.fmt(f)
-            }
-            ComposeError::ParseError(pe) => {
-                write!(f, "parse error in builder: {}", pe)
-            }
-
-        }
-    }
-}
-
-use core::convert::Infallible;
-impl<Target: OctetsBuilder + AsMut<[u8]>> UpdateBuilder<Target>
-where Infallible: From<<Target as OctetsBuilder>::AppendError>
-{
-    pub fn build_acs(mut self, acs: AttrChangeSet)
-        -> Result<Target, ComposeError>
-    {
-        // Withdrawals
-        let withdraw_len = 0_usize;
-        // placeholder
-        let _ = self.target.append_slice(&(withdraw_len as u16).to_be_bytes());
-        //self.target.as_mut()[19..=20].copy_from_slice(
-        //    &(withdraw_len as u16).to_be_bytes()
-        //);
-        // TODO actual withdrawals
-
-
-        // Path Attributes
-            // flags (from msb to lsb):
-            // optional
-            // transitive
-            // partial
-            // extended_length (2 octet length)
-        
-        let mut total_pa_len = 0_usize;
-        // Total Path Attribute len place holder:
-        let _ = self.target.append_slice(&[0x00, 0x00]);
-
-        if let Some(origin) = acs.origin_type.into_opt() {
-            let attr_flags = 0b0100_0000;
-            let attr_typecode = PathAttributeType::Origin.into();
-            let attr_len = 1_u8; 
-            let _ = self.target.append_slice(
-                &[attr_flags, attr_typecode, attr_len, origin.into()]);
-            total_pa_len += 2 + 1 + usize::from(attr_len);
-        }
-
-        if let Some(as_path) = acs.as_path.into_opt() {
-            let attr_flags = 0b0101_0000;
-            let attr_typecode = PathAttributeType::AsPath.into();
-            let asp = as_path.into_inner();
-            let attr_len = asp.len();
-            if u16::try_from(attr_len).is_err() {
-                return Err(ComposeError::AttributeTooLarge(
-                    PathAttributeType::AsPath,
-                    attr_len
-                ));
-            }
-            let _ = self.target.append_slice(&[attr_flags, attr_typecode]);
-            let _ = self.target.append_slice(&(attr_len as u16).to_be_bytes());
-            let _ = self.target.append_slice(&asp);
-
-            total_pa_len += 2 + 2 + attr_len;
-        }
-
-
-        // XXX the next_hop is either a (conventional, for v4/unicast) path
-        // attribute, or, it is part of MP_REACH_NLRI.
-        // Should/must v4/unicast always go in MP_REACH_NLRI when both peers
-        // sent such capability though?
-        if let Some(next_hop) = acs.next_hop.into_opt() {
-            match next_hop {
-                NextHop::Ipv4(v4addr) => {
-                    let attr_flags = 0b0100_0000;
-                    let attr_typecode = PathAttributeType::NextHop.into();
-                    let attr_len = 4_u8; 
-
-                    let _ = self.target.append_slice(
-                        &[attr_flags, attr_typecode, attr_len]
-                    );
-                    let _ = self.target.append_slice(&v4addr.octets());
-
-                    total_pa_len += 2 + 1 + usize::from(attr_len);
-                }
-                _ => todo!() // this is MP_REACH_NLRI territory
-            }
-        }
-
-
-        if let Some(comms) = acs.standard_communities.into_opt() {
-            let attr_flags = 0b0100_0000;
-            let attr_typecode = PathAttributeType::Communities.into();
-            let attr_len = match u8::try_from(4 * comms.len()) {
-                Ok(n) => n,
-                Err(..) => {
-                    return Err(ComposeError::AttributeTooLarge(
-                        PathAttributeType::Communities,
-                        4 * comms.len()
-                    ));
-                }
-            };
-
-            let _ = self.target.append_slice(
-                &[attr_flags, attr_typecode, attr_len]
-            );
-
-            for c in comms {
-                let _ = self.target.append_slice(&c.to_raw());
-            }
-            total_pa_len += 2 + 1 + usize::from(attr_len);
-        }
-
-
-        if u16::try_from(total_pa_len).is_err() {
-            return Err(ComposeError::AttributesTooLarge(total_pa_len));
-        }
-
-        // update total path attribute len:
-        self.target.as_mut()[21+withdraw_len..21+withdraw_len+2]
-            .copy_from_slice(&(total_pa_len as u16).to_be_bytes());
-
-
-        // NLRI
-        // TODO this all needs to be a lot more sophisticated:
-        //  - prefixes can not occur in both withdrawals and nlris, so check
-        //  for that;
-        //  - non v4/unicast NLRI should go in MP_REACH_NLRI, not here (at the
-        //  end of the PDU);
-        //  - we should be able to put multiple NLRI in one UPDATE, though
-        //  currently the AttrChangeSet only holds one;
-        //  - probably more
-        
-        let mut nlri_len = 0;
-
-        if let Some(nlri) = acs.nlri.into_opt() {
-            match nlri {
-                Nlri::Unicast(b) => {
-                    if let Some(id) = b.path_id() {
-                        let _ = self.target.append_slice(&id.to_raw());
-                        nlri_len += 4;
-                    }
-                    match b.prefix().addr_and_len() {
-                        (std::net::IpAddr::V4(addr), len) => {
-                            let _ = self.target.append_slice(&[len]);
-                            let len_bytes = (usize::from(len)-1) / 8 + 1;
-                            let _ = self.target.append_slice(
-                                &addr.octets()[0..len_bytes]
-                            );
-                            nlri_len += 1 + len_bytes;
-                        }
-                        _ => todo!()
-                    }
-                }
-                _ => todo!()
-            }
-        }
-
-        // update pdu len
-        let msg_len = 19 
-            + 2 + withdraw_len 
-            + 2 + total_pa_len
-            + nlri_len
-        ;
-
-        if msg_len > 4096 {
-            // TODO handle Extended Messages (max pdu size 65535)
-            return Err(ComposeError::PduTooLarge(msg_len));
-        }
-
-        //if u16::try_from(msg_len).is_err() {
-        //    return Err(ComposeError::PduTooLarge(msg_len));
-        //}
-
-        self.target.as_mut()[16..=17].copy_from_slice(
-            &(msg_len as u16).to_be_bytes()
-        );
-
-        Ok(self.target)
-    }
-}
-
-use new_pas::Attribute;
-impl<Target> UpdateBuilder<Target>
-where
-    Target: OctetsBuilder + FreezeBuilder + AsMut<[u8]>,
-    <Target as FreezeBuilder>::Octets: Octets,
-{
-    pub fn into_message(self) ->
-        Result<UpdateMessage<<Target as FreezeBuilder>::Octets>, ComposeError>
-    where
-        Target: FreezeBuilder,
-        <Target as FreezeBuilder>::Octets: Octets,
-    {
-        self.is_valid()?;
-        Ok(UpdateMessage::from_octets(
-            self.finish().map_err(|_| ShortBuf)?, SessionConfig::modern()
-        )?)
-    }
-
-    // Check whether the combination of NLRI and attributes would produce a
-    // valid UPDATE pdu.
-    fn is_valid(&self) -> Result<(), ComposeError> {
-        // If we have builders for MP_(UN)REACH_NLRI, they should carry
-        // prefixes.
-        if let Some(ref builder) = self.mp_reach_nlri_builder {
-            if builder.is_empty() {
-                return Err(ComposeError::EmptyMpReachNlri);
-            }
-        }
-        if let Some(ref builder) = self.mp_unreach_nlri_builder {
-            if builder.is_empty() {
-                return Err(ComposeError::EmptyMpUnreachNlri);
-            }
-        }
-
-        Ok(())
-    }
-
-    fn finish(mut self)
-        -> Result<<Target as FreezeBuilder>::Octets, Target::AppendError>
-    {
-        let mut header = Header::for_slice_mut(self.target.as_mut());
-        header.set_length(u16::try_from(self.total_pdu_len).unwrap());
-
-        // `withdrawals_len` is checked to be <= 4096 or <= 65535
-        // so it will always fit in a u16.
-        self.target.append_slice(
-            &u16::try_from(self.withdrawals_len).unwrap().to_be_bytes()
-        )?;
-
-        for w in &self.withdrawals {
-            match w {
-                Nlri::Unicast(b) => {
-                    if b.is_v4() {
-                        b.compose(&mut self.target)?;
-                    } else {
-                        // Other withdrawals should not go here, but in
-                        // MP_UNREACH_NLRI. Handling these ones below in the
-                        // path attributes.
-                    }
-                },
-                _ => todo!(),
-            }
-        }
-
-
-
-        // XXX we can do these unwraps because of the checks in the add/append
-        // methods
-
-        // `attributes_len` is checked to be <= 4096 or <= 65535
-        // so it will always fit in a u16.
-        let _ = self.target.append_slice(
-            &u16::try_from(self.attributes_len).unwrap().to_be_bytes()
-        );
-
-        // TODO write all path attributes, if any, in order of typecode
-
-        if let Some(origin) = self.origin {
-            origin.compose(&mut self.target)?
-        }
-
-        if let Some(aspath) = self.aspath {
-            aspath.compose(&mut self.target)?
-        }
-
-        if let Some(nexthop) = self.nexthop {
-            nexthop.compose(&mut self.target)?
-        }
-
-        if let Some(med) = self.multi_exit_disc {
-            med.compose(&mut self.target)?
-        }
-
-        if let Some(local_pref) = self.local_pref {
-            local_pref.compose(&mut self.target)?
-        }
-
-        if let Some(builder) = self.standard_communities_builder {
-            builder.compose(&mut self.target)?
-        }
-
-        if let Some(builder) = self.mp_reach_nlri_builder {
-            builder.compose(&mut self.target)?
-        }
-
-        if let Some(builder) = self.mp_unreach_nlri_builder {
-            builder.compose(&mut self.target)?
-        }
-
-        // XXX Here, in the conventional NLRI field at the end of the PDU, we
-        // write IPv4 Unicast announcements. But what if we have agreed to do
-        // MP for v4/unicast, should these announcements go in the
-        // MP_REACH_NLRI attribute then instead?
-        for a in &self.announcements {
-            match a {
-                Nlri::Unicast(b) => {
-                    if b.is_v4() {
-                        b.compose(&mut self.target)?;
-                    } else {
-                        // Other announcements should not go here, but in
-                        // MP_REACH_NLRI, handled before in the path
-                        // attributes.
-                    }
-                },
-                _ => todo!(),
-            }
-        }
-
-        Ok(self.target.freeze())
-    }
-}
-
-impl UpdateBuilder<Vec<u8>> {
-    pub fn new_vec() -> Self {
-        Self::from_target(Vec::with_capacity(23)).unwrap()
-    }
-}
-
-impl UpdateBuilder<BytesMut> {
-    pub fn new_bytes() -> Self {
-        Self::from_target(BytesMut::new()).unwrap()
-    }
-}
-
 //--- Tests ------------------------------------------------------------------
 
 #[cfg(test)]
@@ -2910,12 +2044,21 @@ mod tests {
 
     use super::*;
     use std::str::FromStr;
-    use std::net::Ipv6Addr;
     use crate::bgp::communities::*;
     use crate::bgp::message::{Message, nlri::PathId};
     use crate::addr::Prefix;
 
     use bytes::Bytes;
+
+    #[allow(dead_code)]
+    fn print_pcap<T: AsRef<[u8]>>(buf: T) {
+        print!("000000 ");
+        for b in buf.as_ref() {
+            print!("{:02x} ", b);
+        }
+        println!();
+    }
+
 
     //TODO:
     // X generic
@@ -3203,27 +2346,6 @@ mod tests {
             "2001:db8:ffff:3::/64",
         ].map(|w| Nlri::unicast_from_str(w).unwrap());
         assert!(ws.into_iter().eq(update.withdrawals().iter()));
-    }
-
-    #[test]
-    fn empty_nlri_iterators() {
-        let mut builder = UpdateBuilder::new_vec();
-        builder.add_withdrawal(
-            &Nlri::unicast_from_str("2001:db8::/32").unwrap()
-        ).unwrap();
-
-        let msg = builder.into_message().unwrap();
-        print_pcap(msg.as_ref());
-        assert_eq!(msg.all_withdrawals_iter().count(), 1);
-
-        let mut builder2 = UpdateBuilder::new_vec();
-        builder2.add_withdrawal(
-            &Nlri::unicast_from_str("10.0.0.0/8").unwrap()
-        ).unwrap();
-
-        let msg2 = builder2.into_message().unwrap();
-        print_pcap(msg2.as_ref());
-        assert_eq!(msg2.all_withdrawals_iter().count(), 1);
     }
 
     //--- Path Attributes ------------------------------------------------
@@ -3587,539 +2709,6 @@ mod tests {
         assert_eq!(upd.withdrawals().afi(), AFI::Ipv4);
         assert_eq!(upd.withdrawals().safi(), SAFI::Multicast);
         assert_eq!(upd.withdrawals().iter().count(), 1);
-    }
-
-    fn print_pcap<T: AsRef<[u8]>>(buf: T) {
-        print!("000000 ");
-        for b in buf.as_ref() {
-            print!("{:02x} ", b);
-        }
-        println!();
-    }
-
-    #[test]
-    fn build_empty() {
-        let builder = UpdateBuilder::new_vec();
-        let msg = builder.finish().unwrap();
-        //print_pcap(&msg);
-        assert!(
-            UpdateMessage::from_octets(msg, SessionConfig::modern()).is_ok()
-        );
-    }
-
-    #[test]
-    fn build_withdrawals_basic_v4() {
-        let mut builder = UpdateBuilder::new_vec();
-
-        let withdrawals = [
-            "0.0.0.0/0",
-            "10.2.1.0/24",
-            "10.2.2.0/24",
-            "10.2.0.0/23",
-            "10.2.4.0/25",
-            "10.0.0.0/7",
-            "10.0.0.0/8",
-            "10.0.0.0/9",
-        ].map(|s| Nlri::unicast_from_str(s).unwrap())
-         .into_iter()
-         .collect::<Vec<_>>();
-
-        let _ = builder.append_withdrawals(&mut withdrawals.clone());
-        let msg = builder.finish().unwrap();
-        assert!(
-            UpdateMessage::from_octets(&msg, SessionConfig::modern())
-            .is_ok()
-        );
-        print_pcap(&msg);
-
-
-        let mut builder2 = UpdateBuilder::new_vec();
-        for w in &withdrawals {
-            builder2.add_withdrawal(w).unwrap();
-        }
-
-        let msg2 = builder2.finish().unwrap();
-        assert!(
-            UpdateMessage::from_octets(&msg2, SessionConfig::modern())
-            .is_ok()
-        );
-        print_pcap(&msg2);
-
-        assert_eq!(msg, msg2);
-    }
-
-    #[test]
-    fn build_withdrawals_from_iter() {
-        let mut builder = UpdateBuilder::new_vec();
-
-        let withdrawals = [
-            "0.0.0.0/0",
-            "10.2.1.0/24",
-            "10.2.2.0/24",
-            "10.2.0.0/23",
-            "10.2.4.0/25",
-            "10.0.0.0/7",
-            "10.0.0.0/8",
-            "10.0.0.0/9",
-        ].map(|s| Nlri::unicast_from_str(s).unwrap().into())
-         .into_iter()
-         .collect::<Vec<_>>();
-
-        let _ = builder.withdrawals_from_iter(&mut withdrawals.into_iter().peekable());
-        let msg = builder.into_message().unwrap();
-        print_pcap(msg);
-    }
-
-    #[test]
-    #[should_panic]
-    fn build_too_many_withdrawals() {
-        let mut builder = UpdateBuilder::new_vec();
-        for i in 0..1024 {
-            builder.add_withdrawal(
-                &Nlri::unicast_from_str(&format!("2001:db:{:04}::/48", i))
-                    .unwrap()
-            ).unwrap();
-        }
-    }
-
-    #[test]
-    fn build_too_many_withdrawals_remainder() {
-        let mut builder = UpdateBuilder::new_vec();
-        let mut prefixes: Vec<Nlri<Vec<u8>>> = vec![];
-        for i in 1..1024_u32 {
-            prefixes.push(
-                Nlri::Unicast(
-                    Prefix::new_v4(
-                        Ipv4Addr::from((i << 10).to_be_bytes()),
-                        22
-                    ).unwrap().into()
-                )
-            );
-        }
-        let prefixes_len = prefixes.len();
-        assert!(builder.append_withdrawals(&mut prefixes).is_err());
-        assert!(prefixes.len() < prefixes_len);
-        let pdu = builder.into_message().unwrap();
-        assert_eq!(
-            pdu.withdrawals().iter().count(),
-            prefixes_len - prefixes.len()
-        );
-    }
-
-    #[test]
-    fn build_too_many_withdrawals_from_iter() {
-        let mut builder = UpdateBuilder::new_vec();
-        let mut prefixes: Vec<Nlri<Vec<u8>>> = vec![];
-        for i in 1..1024_u32 {
-            prefixes.push(
-                Nlri::Unicast(
-                    Prefix::new_v4(
-                        Ipv4Addr::from((i << 10).to_be_bytes()),
-                        22
-                    ).unwrap().into()
-                )
-            );
-        }
-        let mut prefix_iter = prefixes.into_iter().peekable();
-
-        assert!(builder.withdrawals_from_iter(&mut prefix_iter).is_err());
-        let msg = builder.into_message().unwrap();
-        assert_eq!(
-            prefix_iter.count() + msg.withdrawals().iter().count(),
-            1023
-        );
-    }
-
-
-
-    #[test]
-    fn build_withdrawals_basic_v4_addpath() {
-        use crate::bgp::message::nlri::PathId;
-        let mut builder = UpdateBuilder::new_vec();
-        let mut withdrawals = [
-            "0.0.0.0/0",
-            "10.2.1.0/24",
-            "10.2.2.0/24",
-            "10.2.0.0/23",
-            "10.2.4.0/25",
-            "10.0.0.0/7",
-            "10.0.0.0/8",
-            "10.0.0.0/9",
-        ].iter().enumerate().map(|(idx, s)| Nlri::Unicast(BasicNlri {
-            prefix: Prefix::from_str(s).unwrap(),
-            path_id: Some(PathId::from_u32(idx.try_into().unwrap()))})
-        ).collect::<Vec<_>>();
-        let _ = builder.append_withdrawals(&mut withdrawals);
-        let msg = builder.finish().unwrap();
-        assert!(
-            UpdateMessage::from_octets(&msg, SessionConfig::modern_addpath())
-            .is_ok()
-        );
-        print_pcap(&msg);
-    }
-
-    #[test]
-    fn build_withdrawals_basic_v6() {
-        let mut builder = UpdateBuilder::new_vec();
-        let mut withdrawals = vec![
-            Nlri::unicast_from_str("2001:db8::/32").unwrap()
-        ];
-
-        let _ = builder.append_withdrawals(&mut withdrawals);
-
-        let msg = builder.finish().unwrap();
-        println!("msg raw len: {}", &msg.len());
-        print_pcap(&msg);
-        
-        UpdateMessage::from_octets(&msg, SessionConfig::modern()).unwrap();
-    }
-
-    #[test]
-    fn build_withdrawals_basic_v6_from_iter() {
-        let mut builder = UpdateBuilder::new_vec();
-
-        let mut withdrawals: Vec<Nlri<Vec<u8>>> = vec![];
-        for i in 1..512_u128 {
-            withdrawals.push(
-                Nlri::Unicast(
-                    Prefix::new_v6(
-                        Ipv6Addr::from((i << 64).to_be_bytes()),
-                        64
-                    ).unwrap().into()
-                )
-            );
-        }
-
-        let _ = builder.withdrawals_from_iter(&mut withdrawals.into_iter().peekable());
-        let raw = builder.finish().unwrap();
-        print_pcap(&raw);
-        UpdateMessage::from_octets(&raw, SessionConfig::modern()).unwrap();
-    }
-
-    #[test]
-    fn build_mixed_withdrawals() {
-        let mut builder = UpdateBuilder::new_vec();
-        builder.add_withdrawal(
-            &Nlri::unicast_from_str("10.0.0.0/8").unwrap()
-        ).unwrap();
-        builder.add_withdrawal(
-            &Nlri::unicast_from_str("2001:db8::/32").unwrap()
-        ).unwrap();
-        let msg = builder.into_message().unwrap();
-        print_pcap(msg.as_ref());
-
-        assert_eq!(msg.all_withdrawals_iter().count(), 2);
-    }
-
-    #[test]
-    fn build_mixed_addpath_conventional() {
-        use crate::bgp::message::nlri::PathId;
-        let mut builder = UpdateBuilder::new_vec();
-        builder.add_withdrawal(
-            &Nlri::unicast_from_str("10.0.0.0/8").unwrap()
-        ).unwrap();
-        let res = builder.add_withdrawal(
-            &Nlri::Unicast(
-                (Prefix::from_str("10.0.0.0/8").unwrap(),
-                Some(PathId::from_u32(123))
-                ).into())
-        );
-        assert!(matches!(res, Err(ComposeError::IllegalCombination)));
-    }
-
-    #[test]
-    fn build_mixed_addpath_mp() {
-        use crate::bgp::message::nlri::PathId;
-        let mut builder = UpdateBuilder::new_vec();
-        builder.add_withdrawal(
-            &Nlri::unicast_from_str("2001:db8::/32").unwrap()
-        ).unwrap();
-        let res = builder.add_withdrawal(
-            &Nlri::Unicast(
-                (Prefix::from_str("2001:db8::/32").unwrap(),
-                Some(PathId::from_u32(123))
-                ).into())
-        );
-        assert!(matches!(res, Err(ComposeError::IllegalCombination)));
-    }
-
-    #[test]
-    fn build_mixed_addpath_ok() {
-        use crate::bgp::message::nlri::PathId;
-        let mut builder = UpdateBuilder::new_vec();
-        builder.add_withdrawal(
-            &Nlri::unicast_from_str("10.0.0.0/8").unwrap()
-        ).unwrap();
-        let res = builder.add_withdrawal(
-            &Nlri::Unicast(
-                (Prefix::from_str("2001:db8::/32").unwrap(),
-                Some(PathId::from_u32(123))
-                ).into())
-        );
-        assert!(res.is_ok());
-        print_pcap(builder.finish().unwrap());
-    }
-
-    #[test]
-    fn build_conv_mp_mix() {
-        let buf = vec![
-            0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff,
-            0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff,
-            0x00, 0x88 + 8, // adding length for the conv NLRI
-            0x02, 0x00, 0x00, 0x00, 0x71, 0x80,
-            0x0e, 0x5a, 0x00, 0x02, 0x01, 0x20, 0xfc, 0x00,
-            0x00, 0x10, 0x00, 0x01, 0x00, 0x10, 0x00, 0x00,
-            0x00, 0x00, 0x00, 0x00, 0x00, 0x10, 0xfe, 0x80,
-            0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-            0x00, 0x00, 0x00, 0x00, 0x00, 0x10, 0x00, 0x80,
-            0xfc, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-            0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x10,
-            0x40, 0x20, 0x01, 0x0d, 0xb8, 0xff, 0xff, 0x00,
-            0x00, 0x40, 0x20, 0x01, 0x0d, 0xb8, 0xff, 0xff,
-            0x00, 0x01, 0x40, 0x20, 0x01, 0x0d, 0xb8, 0xff,
-            0xff, 0x00, 0x02, 0x40, 0x20, 0x01, 0x0d, 0xb8,
-            0xff, 0xff, 0x00, 0x03, 0x40, 0x01, 0x01, 0x00,
-            0x40, 0x02, 0x06, 0x02, 0x01, 0x00, 0x00, 0x00,
-            0xc8, 0x80, 0x04, 0x04, 0x00, 0x00, 0x00, 0x00,
-            // manually adding two conv NLRI here
-            24, 1, 1, 1,
-            24, 1, 1, 2
-
-        ];
-
-        let upd = UpdateMessage::from_octets(&buf, SessionConfig::modern()).unwrap();
-        print_pcap(upd.as_ref());
-
-        assert!(upd.has_conventional_nlri() && upd.has_mp_nlri());
-        assert_eq!(upd.unicast_announcements().count(), 7);
-    }
-
-    #[test]
-    fn build_announcements_conventional() {
-        use crate::bgp::aspath::HopPath;
-        let mut builder = UpdateBuilder::new_vec();
-        let prefixes = [
-            "1.0.1.0/24",
-            "1.0.2.0/24",
-            "1.0.3.0/24",
-            "1.0.4.0/24",
-        ].map(|p| Nlri::unicast_from_str(p).unwrap());
-        let mut iter = prefixes.into_iter().peekable();
-        builder.announcements_from_iter(&mut iter).unwrap();
-        builder.set_origin(OriginType::Igp).unwrap();
-        builder.set_nexthop("1.2.3.4".parse().unwrap()).unwrap();
-        let path = HopPath::from([
-             Asn::from_u32(123); 70
-        ]);
-        builder.set_aspath::<Vec<u8>>(path.to_as_path().unwrap()).unwrap();
-
-        let raw = builder.finish().unwrap();
-        print_pcap(&raw);
-
-        //let pdu = builder.into_message().unwrap();
-        //print_pcap(pdu);
-    }
-
-    #[test]
-    fn build_announcements_mp() {
-        use crate::bgp::aspath::HopPath;
-
-        let mut builder = UpdateBuilder::new_vec();
-        let prefixes = [
-            "2001:db8:1::/48",
-            "2001:db8:2::/48",
-            "2001:db8:3::/48",
-        ].map(|p| Nlri::unicast_from_str(p).unwrap());
-        let mut iter = prefixes.into_iter().peekable();
-        builder.announcements_from_iter(&mut iter).unwrap();
-        builder.set_origin(OriginType::Igp).unwrap();
-        builder.set_nexthop("fe80:1:2:3::".parse().unwrap()).unwrap();
-        let path = HopPath::from([
-             Asn::from_u32(100),
-             Asn::from_u32(200),
-             Asn::from_u32(300),
-        ]);
-        builder.set_aspath::<Vec<u8>>(path.to_as_path().unwrap()).unwrap();
-
-        let raw = builder.finish().unwrap();
-        print_pcap(&raw);
-    }
-
-    #[test]
-    fn build_announcements_mp_missing_nlri() {
-        use crate::bgp::aspath::HopPath;
-
-        let mut builder = UpdateBuilder::new_vec();
-        builder.set_origin(OriginType::Igp).unwrap();
-        builder.set_nexthop("fe80:1:2:3::".parse().unwrap()).unwrap();
-        assert!(builder.mp_reach_nlri_builder.is_some());
-        let path = HopPath::from([
-             Asn::from_u32(100),
-             Asn::from_u32(200),
-             Asn::from_u32(300),
-        ]);
-        builder.set_aspath::<Vec<u8>>(path.to_as_path().unwrap()).unwrap();
-
-        assert!(matches!(
-                builder.into_message(),
-                Err(ComposeError::EmptyMpReachNlri)
-        ));
-    }
-
-    #[test]
-    fn build_announcements_mp_link_local() {
-        use crate::bgp::aspath::HopPath;
-
-        let mut builder = UpdateBuilder::new_vec();
-
-        let prefixes = [
-            "2001:db8:1::/48",
-            "2001:db8:2::/48",
-            "2001:db8:3::/48",
-        ].map(|p| Nlri::unicast_from_str(p).unwrap());
-
-
-        let mut iter = prefixes.into_iter().peekable();
-
-        builder.announcements_from_iter(&mut iter).unwrap();
-        builder.set_origin(OriginType::Igp).unwrap();
-        //builder.set_nexthop("2001:db8::1".parse().unwrap()).unwrap();
-        builder.set_nexthop_ll("fe80:1:2:3::".parse().unwrap()).unwrap();
-
-
-        let path = HopPath::from([
-             Asn::from_u32(100),
-             Asn::from_u32(200),
-             Asn::from_u32(300),
-        ]);
-        builder.set_aspath::<Vec<u8>>(path.to_as_path().unwrap()).unwrap();
-
-        //let unchecked = builder.finish().unwrap();
-        //print_pcap(unchecked);
-        let msg = builder.into_message().unwrap();
-        msg.print_pcap();
-    }
-
-    #[test]
-    fn build_announcements_mp_ll_no_nlri() {
-        use crate::bgp::aspath::HopPath;
-
-        let mut builder = UpdateBuilder::new_vec();
-        builder.set_origin(OriginType::Igp).unwrap();
-        //builder.set_nexthop("2001:db8::1".parse().unwrap()).unwrap();
-        builder.set_nexthop_ll("fe80:1:2:3::".parse().unwrap()).unwrap();
-        assert!(builder.mp_reach_nlri_builder.is_some());
-        let path = HopPath::from([
-             Asn::from_u32(100),
-             Asn::from_u32(200),
-             Asn::from_u32(300),
-        ]);
-        builder.set_aspath::<Vec<u8>>(path.to_as_path().unwrap()).unwrap();
-
-        assert!(matches!(
-                builder.into_message(),
-                Err(ComposeError::EmptyMpReachNlri)
-        ));
-    }
-
-    #[test]
-    fn build_standard_communities() {
-        use crate::bgp::aspath::HopPath;
-        let mut builder = UpdateBuilder::new_vec();
-        let prefixes = [
-            "1.0.1.0/24",
-            "1.0.2.0/24",
-            "1.0.3.0/24",
-            "1.0.4.0/24",
-        ].map(|p| Nlri::unicast_from_str(p).unwrap());
-        let mut iter = prefixes.into_iter().peekable();
-        builder.announcements_from_iter(&mut iter).unwrap();
-        builder.set_origin(OriginType::Igp).unwrap();
-        builder.set_nexthop("1.2.3.4".parse().unwrap()).unwrap();
-        let path = HopPath::from([
-             Asn::from_u32(100),
-             Asn::from_u32(200),
-             Asn::from_u32(300),
-        ]);
-        builder.set_aspath::<Vec<u8>>(path.to_as_path().unwrap()).unwrap();
-
-
-        builder.add_community("AS1234:666".parse().unwrap()).unwrap();
-        builder.add_community("NO_EXPORT".parse().unwrap()).unwrap();
-        for n in 1..100 {
-            builder.add_community(format!("AS999:{n}").parse().unwrap()).unwrap();
-        }
-
-        builder.into_message().unwrap();
-        //let raw = builder.finish().unwrap();
-        //print_pcap(&raw);
-    }
-
-    #[test]
-    fn build_other_attributes() {
-        use crate::bgp::aspath::HopPath;
-        let mut builder = UpdateBuilder::new_vec();
-        let prefixes = [
-            "1.0.1.0/24",
-            "1.0.2.0/24",
-            "1.0.3.0/24",
-            "1.0.4.0/24",
-        ].map(|p| Nlri::unicast_from_str(p).unwrap());
-        let mut iter = prefixes.into_iter().peekable();
-        builder.announcements_from_iter(&mut iter).unwrap();
-        builder.set_origin(OriginType::Igp).unwrap();
-        builder.set_nexthop("1.2.3.4".parse().unwrap()).unwrap();
-        let path = HopPath::from([
-             Asn::from_u32(100),
-             Asn::from_u32(200),
-             Asn::from_u32(300),
-        ]);
-        builder.set_aspath::<Vec<u8>>(path.to_as_path().unwrap()).unwrap();
-
-        builder.set_multi_exit_disc(new_pas::MultiExitDisc::new(1234)).unwrap();
-        builder.set_local_pref(new_pas::LocalPref::new(9876)).unwrap();
-
-        let msg = builder.into_message().unwrap();
-        msg.print_pcap();
-    }
-
-
-
-    #[test]
-    fn build_acs() {
-        use crate::bgp::aspath::HopPath;
-        use crate::bgp::message::nlri::PathId;
-
-        let builder = UpdateBuilder::new_vec();
-        let mut acs = AttrChangeSet::empty();
-
-        // ORIGIN
-        acs.origin_type.set(OriginType::Igp);
-
-        // AS_PATH
-        let mut hp = HopPath::new();
-        hp.prepend(Asn::from_u32(100));
-        hp.prepend(Asn::from_u32(101));
-        acs.as_path.set(hp.to_as_path().unwrap());
-
-        // NEXT_HOP
-        acs.next_hop.set(NextHop::Ipv4(Ipv4Addr::from_str("192.0.2.1").unwrap()));
-
-
-        // now for some NLRI
-        // XXX currently ACS only holds one single Nlri
-        acs.nlri.set(Nlri::Unicast(BasicNlri{
-            prefix: Prefix::from_str("1.2.0.0/25").unwrap(),
-            path_id: Some(PathId::from_u32(123))
-        }));
-
-        acs.standard_communities.set(vec![
-            Wellknown::NoExport.into(),
-            Wellknown::Blackhole.into(),
-        ]);
-
-        let _msg = builder.build_acs(acs).unwrap();
-        //print_pcap(&msg);
     }
 
 }
